fix: Refactor Streamlit app out into components for modularity
